# dist/openapi.yaml
openapi: 3.0.0
info:
  title: Simple Inventory API
  description: This is a simple API
  contact:
    email: yucoh@post.bgu.ac.il
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: "http://127.0.0.1:3000"
  - url: https://virtserver.swaggerhub.com/yucoh/simple-inventory_api/1.0.0
    description: An API that allows regiser new users to thw website
...
paths:
  /Register:
    post:
      tags:
      - auth
      summary: A new user from client is added to the server
      description: A JSON object contains all the details about the new user
      requestBody:
        description: Registration form filled
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Register'
        required: true
      responses:
        "201":
          description: A new user has been added
        "401":
          description: Wrong input parameters
        "409":
          description: Username was taken
  /Login:
    post:
      tags:
      - auth
      summary: A registered and unconnected user is login to the server
      description: A JSON object contains an exist username with its correct password
      requestBody:
        description: Login form filled
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
        required: true
      responses:
        "200":
          description: The user's login was succeeded
          headers:
            Set-Cookie:
              style: simple
              explode: false
              schema:
                type: string
                example: session=Yuvi
        "401":
          description: Username or password incorrect
  # /Logout:
  #   post:
  #     tags:
  #     - auth
  #     summary: A registered and connected user is logout from the server
  #     description: A JSON object contains an exist username
  #     requestBody:
  #       description: Inventory item to add
  #       required: true
  #     responses:
  #       "201":
  #         description: The user's logout was succeeded
  /user/favorites:
    get:
      tags:
      - user
      summary: Recieve a recipe id and save it to the favorites list of the logged-in user
      description: A JSON object contains an exist username and a recipe id
      responses:
        "201":
          description: The user got his favorites recipes successfully
        "400":
          description: Wrong input parameters
    post:
      tags:
      - user
      summary: Recieve a recipe id and save it to the favorites list of the logged-in user
      description: A JSON object contains an exist username and a recipe id
      requestBody:
        description: Inventory item to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/favorites'
        required: true
      responses:
        "201":
          description: The favorite recipe was added to the user's favorites recipes list successfully
        "400":
          description: Wrong input parameters
  /user/myRecipies:
    get:
      tags:
      - user
      summary: Recieve a recipe id and return it from the server
      description: A JSON object contains an exist recipe id
      responses:
        "201":
          description: The user got the required recipe
        "400":
          description: Wrong input parameters
    post:
      tags:
      - user
      summary: Create a new recipe by a logged-in user
      description: A JSON object contains the all parameters for a new recipe
      requestBody:
        description: Inventory item to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/createRecipe'
        required: true
      responses:
        "201":
          description: A new recipe was added
        "400":
          description: Wrong input parameters
...
components:
  schemas:
    Login:
      required:
        - unser_name
        - password
      type: object
      properties:
        user_name:
          type: string
          example: Yuvi
        password:
          type: string
          example: yuval5210
    createRecipe:
      required:
        - title
        - imageUrl
      type: object
      properties:
        title:
          minLength: 3
          type: string
          description: Recipe name contains at least 3
          example: Chocolate Cake
        imageUrl:
          type: string
          description: URL for the user's image
          format: URL
          example: https://cloundinary.com/chocolateCake
        readyInMinutes:
          type: integer
          description: How long does it take to prepare the recipe
          format: date-time
        popularity:
          type: integer
          description: How popular the recipe is
          example: 20
        vegan:
          type: integer
          description: Is the recipe vegan? Y/N
          example: 0
        vegetarian:
          type: integer
          description: Is the recipe vegetarian? Y/N
          example: 0
        gluten_free:
          type: integer
          description: Does the recipe contain gluten? Y/N
          example: 0
        ingredients:
          minItems: 1
          type: array
          description: Ingredients list of the recipe
          example: Milk 1
          items:
            $ref: '#/components/schemas/createRecipe_ingredients'
        instructions:
          type: string
          description: The preperation instructions of the recipe
          example: First preheat the oven to 165 degrees and grease a 22cm diameter mold...
        servings:
          type: integer
          description: Final number of dishes of the recipe
          format: int32
          example: 3
      description: Model containing recipe info
    Register:
      type: object
      properties:
        username:
          maxLength: 8
          minLength: 3
          uniqueItems: true
          type: string
          description: User name contains between 3 to 8 letters only
          example: yuval
        firstname:
          type: string
          description: User's first name
          example: Yuval
        lastname:
          type: string
          description: User's last name
          example: Cohen
        country:
          type: string
          description: User's country
          example: Israel
        password:
          maxLength: 10
          minLength: 5
          type: string
          description: |
            Password contains between 5 to 10 chars,  at least one of them is a special char and a number
          format: password
          example: yuval!1
        email:
          type: string
          description: User's email
          format: email
          example: yuval@gmail.com
        img-url:
          type: string
          description: URL for the user's image
          format: URL
          example: https://cloundinary.com/yuval
        recipesViewed:
          type: array
          items:
            $ref: '#/components/schemas/createRecipe'
        favoritRecipes:
          type: array
          items:
            $ref: '#/components/schemas/createRecipe'
      description: Model containing user info
    createRecipe_ingredients:
      required:
      - ingredientName
      - ingredientQuantity
      type: object
      properties:
        ingredientName:
          type: string
        ingredientQuantity:
          type: integer
          format: int32
    # Search:
    #   requrired:
    #     - number
    #   type: object
    #   properties:
    #     number:
    #       type: integer
    #       example: 5
    #     query:
    #       type: string
    #       example: chocolate cake
    #     cuisine:
    #       type: string
    #       exapmle: italian
    #     diet:
    #       type: string
    #       example: vegan
    #     intolerance:
    #       type: string
    #       example: gluten
    favorites:
      required: 
        - user_name
        - recipe_id
      type: object
      properties:
        recipe_id:
          type: integer
        user_name:
          type: string

        # username:
        #   maxLength: 8
        #   minLength: 3
        #   uniqueItems: true
        #   type: string
        #   description: User name contains between 3 to 8 letters only
        #   example: yuval
        # firstname:
        #   type: string
        #   description: User's first name
        #   example: Yuval
        # lastname:
        #   type: string
        #   description: User's last name
        #   example: Cohen
        # country:
        #   type: string
        #   description: User's country
        #   example: Israel
        # password:
        #   maxLength: 10
        #   minLength: 5
        #   type: string
        #   description: |
        #     Password contains between 5 to 10 chars,  at least one of them is a special char and a number
        #   format: password
        #   example: yuval!1
        # email:
        #   type: string
        #   description: User's email
        #   format: email
        #   example: yuval@gmail.com
        # img-url:
        #   type: string
        #   description: URL for the user's image
        #   format: URL
        #   example: https://cloundinary.com/yuval
        # recipesViewed:
        #   type: array
        #   items:
        #     $ref: '#/components/schemas/createRecipe'
        # favoritRecipes:
        #   type: array
        #   items:
        #     $ref: '#/components/schemas/createRecipe'
      description: Model containing user info
...